refactor: Move 'screens' folder into 'views'.
