+++ src/views/screens/SignUp.tsx
@@ -17,19 +17,19 @@ import { useForm, Controller } from 'react-hook-form';
 import * as yup from 'yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import LogoSvg from '@assets/logo.svg';
-import { Input } from '@components/Input';
-import { Button } from '@components/Button';
-import { useState } from 'react';
-import { UserPhoto } from '@components/UserPhoto';
-import defaultUserPhotoImg from '@assets/userPhotoDefault.png';
-import { PencilSimpleLine, Eye, EyeSlash } from 'phosphor-react-native';
-import { toMaskedPhone } from '@utils/Masks';
-import * as ImagePicker from 'expo-image-picker';
-import * as FileSystem from 'expo-file-system';
-import { AppError } from '@utils/AppError';
-import uuid from 'react-native-uuid';
-import { api } from '@services/api';
-import { useAuth } from '@hooks/useAuth';
+// import { Input } from '@components/Input';
+// import { Button } from '@components/Button';
+// import { useState } from 'react';
+// import { UserPhoto } from '@components/UserPhoto';
+// import defaultUserPhotoImg from '@assets/userPhotoDefault.png';
+// import { PencilSimpleLine, Eye, EyeSlash } from 'phosphor-react-native';
+// import { toMaskedPhone } from '@utils/Masks';
+// import * as ImagePicker from 'expo-image-picker';
+// import * as FileSystem from 'expo-file-system';
+// import { AppError } from '@utils/AppError';
+// import uuid from 'react-native-uuid';
+// import { api } from '@services/api';
+// import { useAuth } from '@hooks/useAuthViewModel';
 
 const signInSchema = yup.object({
   name: yup.string().required('Informe o nome.'),
@@ -70,102 +70,102 @@ export function SignUp() {
     resolver: yupResolver(signInSchema),
   });
 
-  const { signIn } = useAuth();
+  // const { signIn } = useAuth();
 
-  const [photo, setPhoto] = useState<PhotoProps>({} as PhotoProps);
-  const [photoIsLoading, setPhotoIsLoading] = useState(false);
-  const [passwordSecureTextEntry, setPasswordSecureTextEntry] = useState(true);
-  const [passwordConfirmSecureTextEntry, setPasswordConfirmSecureTextEntry] =
-    useState(true);
-  const [isLoading, setIsLoading] = useState(false);
+  // const [photo, setPhoto] = useState<PhotoProps>({} as PhotoProps);
+  // const [photoIsLoading, setPhotoIsLoading] = useState(false);
+  // const [passwordSecureTextEntry, setPasswordSecureTextEntry] = useState(true);
+  // const [passwordConfirmSecureTextEntry, setPasswordConfirmSecureTextEntry] =
+  //   useState(true);
+  // const [isLoading, setIsLoading] = useState(false);
 
-  function handleGoBack() {
-    navigation.goBack();
-  }
+  // function handleGoBack() {
+  //   navigation.goBack();
+  // }
 
-  async function handleSignUp({ name, password, email, phone }: FormDataProps) {
-    try {
-      const formData = new FormData();
-      formData.append('name', name);
-      formData.append('email', email);
-      formData.append('tel', phone);
-      formData.append('password', password);
-      if (!!photo.uri) {
-        formData.append('avatar', photo as any);
-      }
-      setIsLoading(true);
-      await api.post('/users', formData, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-        },
-      });
-      await signIn(email, password);
-    } catch (error) {
-      const isAppError = error instanceof AppError;
-      const title = isAppError
-        ? error.message
-        : 'Não foi possível criar a conta. Tente novamente mais tarde.';
+  // async function handleSignUp({ name, password, email, phone }: FormDataProps) {
+  //   try {
+  //     const formData = new FormData();
+  //     formData.append('name', name);
+  //     formData.append('email', email);
+  //     formData.append('tel', phone);
+  //     formData.append('password', password);
+  //     if (!!photo.uri) {
+  //       formData.append('avatar', photo as any);
+  //     }
+  //     setIsLoading(true);
+  //     await api.post('/users', formData, {
+  //       headers: {
+  //         'Content-Type': 'multipart/form-data',
+  //       },
+  //     });
+  //     await signIn(email, password);
+  //   } catch (error) {
+  //     const isAppError = error instanceof AppError;
+  //     const title = isAppError
+  //       ? error.message
+  //       : 'Não foi possível criar a conta. Tente novamente mais tarde.';
 
-      toast.show({
-        title,
-        placement: 'top',
-        bgColor: 'red.500',
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  }
+  //     toast.show({
+  //       title,
+  //       placement: 'top',
+  //       bgColor: 'red.500',
+  //     });
+  //   } finally {
+  //     setIsLoading(false);
+  //   }
+  // }
 
-  async function handleUserPhotoSelect() {
-    setPhotoIsLoading(true);
-    try {
-      const photoSelected = await ImagePicker.launchImageLibraryAsync({
-        mediaTypes: ImagePicker.MediaTypeOptions.Images,
-        quality: 1,
-        aspect: [4, 4],
-        allowsEditing: true,
-      });
+  // async function handleUserPhotoSelect() {
+  //   setPhotoIsLoading(true);
+  //   try {
+  //     const photoSelected = await ImagePicker.launchImageLibraryAsync({
+  //       mediaTypes: ImagePicker.MediaTypeOptions.Images,
+  //       quality: 1,
+  //       aspect: [4, 4],
+  //       allowsEditing: true,
+  //     });
 
-      if (photoSelected.canceled) return;
+  //     if (photoSelected.canceled) return;
 
-      if (photoSelected.assets[0].uri) {
-        const photoInfo = (await FileSystem.getInfoAsync(
-          photoSelected.assets[0].uri
-        )) as FileSystem.FileInfo & { size?: number };
+  //     if (photoSelected.assets[0].uri) {
+  //       const photoInfo = (await FileSystem.getInfoAsync(
+  //         photoSelected.assets[0].uri
+  //       )) as FileSystem.FileInfo & { size?: number };
 
-        if (photoInfo.size && photoInfo.size / 1024 / 1024 > 5) {
-          setPhoto({} as PhotoProps);
-          return toast.show({
-            title: 'Essa Imagem é muito grande. Escolha uma de até 5MB',
-            placement: 'top',
-            bgColor: 'red.500',
-          });
-        }
+  //       if (photoInfo.size && photoInfo.size / 1024 / 1024 > 5) {
+  //         setPhoto({} as PhotoProps);
+  //         return toast.show({
+  //           title: 'Essa Imagem é muito grande. Escolha uma de até 5MB',
+  //           placement: 'top',
+  //           bgColor: 'red.500',
+  //         });
+  //       }
 
-        const fileExtension = photoSelected.assets[0].uri.split('.').pop();
+  //       const fileExtension = photoSelected.assets[0].uri.split('.').pop();
 
-        const photoFile = {
-          name: `${String(uuid.v4())}.${fileExtension}`.toLowerCase(),
-          uri: photoSelected.assets[0].uri,
-          type: `${photoSelected.assets[0].type}/${fileExtension}`,
-        } as any;
+  //       const photoFile = {
+  //         name: `${String(uuid.v4())}.${fileExtension}`.toLowerCase(),
+  //         uri: photoSelected.assets[0].uri,
+  //         type: `${photoSelected.assets[0].type}/${fileExtension}`,
+  //       } as any;
 
-        setPhoto(photoFile);
-      }
-    } catch (error) {
-      const isAppError = error instanceof AppError;
-      const title = isAppError
-        ? error.message
-        : 'Não foi possível utilizar a foto. Tente novamente mais tarde.';
-      toast.show({
-        title,
-        placement: 'top',
-        bgColor: 'red.500',
-      });
-    } finally {
-      setPhotoIsLoading(false);
-    }
-  }
+  //       setPhoto(photoFile);
+  //     }
+  //   } catch (error) {
+  //     const isAppError = error instanceof AppError;
+  //     const title = isAppError
+  //       ? error.message
+  //       : 'Não foi possível utilizar a foto. Tente novamente mais tarde.';
+  //     toast.show({
+  //       title,
+  //       placement: 'top',
+  //       bgColor: 'red.500',
+  //     });
+  //   } finally {
+  //     setPhotoIsLoading(false);
+  //   }
+  // }
 
   return (
     <KeyboardAvoidingView
@@ -195,7 +195,7 @@ export function SignUp() {
             </Text>
           </Center>
 
-          <Center flex={1}>
+          {/* <Center flex={1}>
             <Box mb={6}>
               {photoIsLoading ? (
                 <Skeleton
@@ -366,7 +366,7 @@ export function SignUp() {
               bgColor='gray.300'
               onPress={handleGoBack}
             />
-          </Center>
+          </Center> */}
         </VStack>
       </ScrollView>
     </KeyboardAvoidingView>
